Add explicit `faction` fields
The existing implementation had a slightly odd mismatch between the
`party` and `party_wikidata` fields, where `party_wikidata` was really
the *faction* wikidata ID.

Instead, keep the `party` field as relating to the party, but add new
`faction` and `faction_wikidata` fields.

Doing this as a Decorator was a little nastier than I had expected, so
switch back to a simple lookup approach.